+++ src/routes/items/+page.svelte
@@ -6,7 +6,7 @@
   import CreateContentDialog from '$lib/features/dialog/items/CreateContentDialog.svelte';
   import CreateFolderDialog from '$lib/features/dialog/items/CreateFolderDialog.svelte';
   import RenameDialog from '$lib/features/dialog/RenameDialog.svelte';
-  import YtImportDialog from '$lib/features/dialog/YtImportDialog.svelte';
+  import YtImportDialog from '$lib/features/youtube/YtImportDialog.svelte';
   import { HierarchyContent, HierarchyFolder } from '$lib/features/hierarchy/items';
   import { setSlots } from '$lib/stores/index.svelte';
   import { current_folder_ids, folder_nodes } from '$lib/stores/items';
